pyproject.toml
[tool.poetry]
name = "dhos-observations-api"
version = "0.0.1"
description = "DHOS Observations API"
authors = ["Adam Romano <adam.romano@sensynehealth.com>"]
keywords = ["Swagger", "Observations"]
repository = "https://github.com/draysontechnologies/dhos-observations-api"

[[tool.poetry.source]]
name = "sensynehealth"
url = "https://pypi.fury.io/sensynehealth/"

[tool.poetry.dependencies]
python = "^3.9"
flask-batteries-included = {version="3.*", extras = ["pgsql", "apispec"]}
kombu-batteries-included = "1.*"
she-logging = "1.*"
python-dateutil = "*"

[tool.poetry.dev-dependencies]
bandit = "*"
black = "*"
coloredlogs = "*"
coverage = "*"
isort = "*"
mock = "*"
mypy = "*"
pytest = "*"
pytest-dhos = {version = "*", extras=["fbi"]}
pytest-flask = "*"
pytest-freezegun = "*"
pytest-mock = "*"
requests_mock = "*"
sadisplay = "*"
safety = "*"
sqlalchemy-stubs = "*"
tox = "*"
tox-docker = "*"
types-python-dateutil = "*"
types-mock = "*"
types-PyYAML ="*"
types-waitress = "*"

[tool.mypy]
namespace_packages=true
ignore_missing_imports=false
disallow_untyped_defs=true

[[tool.mypy.overrides]]
module = [
    "waitress",
    "pytest",
    "flask.testing",
    "apispec.*",
    "connexion",
    "apispec_webframeworks.*",
    "pytest_mock",
    "jose",
    "flask",
    "sadisplay"
]
ignore_missing_imports = true

[tool.isort]
profile = "black"
known_third_party = ["alembic", "apispec", "apispec_webframeworks", "assertpy", "behave", "click", "clients", "connexion", "dateutil", "environs", "flask", "flask_batteries_included", "helpers", "jose", "kombu", "kombu_batteries_included", "marshmallow", "messaging_steps", "mock", "pytest", "pytest_mock", "reporting", "reportportal_behave", "request_steps", "requests", "sadisplay", "she_logging", "sqlalchemy", "waitress", "yaml"]

[tool.black]
line-length = 88
target-version = ["py39"]
